requirements.txt
requests
google-cloud-secret-manager
google-cloud-pubsub